career-flow-ai/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
google-generativeai
supabase
pypdf
tavily-python
pydantic
pinecone
